Revert "iOS13 Test"
This reverts commit cb11f880fdd88f131ca72afbe405f4e1b23c3308.
