+++ Samples/iOS-Swift/iOS-Swift.xcodeproj/xcshareddata/xcschemes/iOS-Swift.xcscheme
@@ -33,8 +33,8 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "7B64386726A6C544000D0F65"
-               BuildableName = "iOS-SwiftTests.xctest"
-               BlueprintName = "iOS-SwiftTests"
+               BuildableName = "iOS-SwiftUITests.xctest"
+               BlueprintName = "iOS-SwiftUITests"
                ReferencedContainer = "container:iOS-Swift.xcodeproj">
             </BuildableReference>
          </TestableReference>
