+++ Samples/iOS-Swift/iOS-SwiftUITests/Info.plist
